Add base force class for simpler model inheritance
